I init getSalesConfig lib.
