+++ hooks/useZoraCreate.tsx
@@ -2,16 +2,13 @@
 
 import { useAccount, usePublicClient } from 'wagmi'
 import { useWriteContracts } from 'wagmi/experimental'
-import {
-  createCreatorClient,
-  FixedPriceParamsType,
-  TimedSaleParamsType,
-} from '@zoralabs/protocol-sdk'
+import { createCreatorClient } from '@zoralabs/protocol-sdk'
 import { CHAIN_ID, REFERRAL_RECIPIENT } from '@/lib/consts'
 import { usePaymasterProvider } from '@/providers/PaymasterProvider'
 import useCreateSuccessRedirect from './useCreateSuccessRedirect'
 import useConnectWallet from './useConnectWallet'
 import { useState } from 'react'
+import getSalesConfig from '@/lib/zora/getSalesConfig'
 
 const useZoraCreate = () => {
   const publicClient = usePublicClient()!
@@ -27,20 +24,7 @@ const useZoraCreate = () => {
       if (!address) await connectWallet()
       const creatorClient = createCreatorClient({ chainId: CHAIN_ID, publicClient })
       const cc0MusicIpfsHash = 'ipfs://bafkreiazqdg6qc3j6yjcxyhvoyaspmjjwal5wvywjs66jobb3pbzknvzxu'
-      console.log('SWEETS saleStrategy', saleStrategy)
-      const timedSaleConfig = {
-        type: 'timed',
-        erc20Name: 'CC0 Music',
-        erc20Symbol: 'CC0',
-      } as TimedSaleParamsType
-      const fixedPriceSaleConfig = {
-        type: 'fixedPrice',
-        pricePerToken: BigInt(1),
-      } as FixedPriceParamsType
-      const salesConfig =
-        saleStrategy === 'ZoraTimedSaleStrategy' ? timedSaleConfig : fixedPriceSaleConfig
-      console.log('SWEETS salesConfig', salesConfig)
-
+      const salesConfig = getSalesConfig(saleStrategy)
       const { parameters } = await creatorClient.create1155({
         contract: {
           name: 'CC0 Music',
